refactor(mcp): isolate guide block state
